refactor(checker): split parameter_checker tests out to hold the 2000-line ceiling
The class-context fix pushed `checkers/parameter_checker.rs` to 2012 physical
lines, over the architecture ceiling. Per the contract this is split rather than
allowlisted: the four in-file `#[cfg(test)]` modules move verbatim to
`src/tests/parameter_checker_tests.rs` (registered from `lib.rs` with its own
`#[cfg(test)]`), taking the module to 1541 lines. `is_jsdoc_param_optional`
widens to `pub(crate)` because the extracted tests call it directly.

Also corrects four rows of the new suite. They asserted "TS1213 exactly once" on
sources that *reference* the offending parameter, and tsc reports the reserved
word twice there — once at the declaration, once at the use. Verified against
typescript@7.0.2: `class H { p = (yield) => yield; }` is `1213,1213,7006` while
`class H { p = (yield) => 0; }` is `1213,7006`. The test sources now match the
oracle-measured fixtures; tsz already agreed with tsc in both shapes.
